# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/v1/auth/access-token": {
            "get": {
                "description": "refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "refresh_token"
                ],
                "summary": "refresh token",
                "parameters": [
                    {
                        "description": "JSON Data",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.AccessTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CommonResponse-auth_SecurityAuthenticatedUser"
                        }
                    }
                }
            }
        },
...
        "/v1/auth/signin": {
            "get": {
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "login godoc",
                "parameters": [
                    {
                        "description": "JSON Data",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CommonResponse-auth_SecurityAuthenticatedUser"
                        }
                    }
                }
            }
        },
        "/v1/auth/signup": {
            "get": {
                "description": "register new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "register user"
                ],
                "summary": "register new user",
                "parameters": [
                    {
                        "description": "JSON Data",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.CommonResponse-auth_SecurityRegisterUser"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.AccessTokenRequest": {
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "auth.DataSecurityAuthenticated": {
            "type": "object",
            "properties": {
                "expirationAccessDateTime": {
                    "type": "string"
                },
                "expirationRefreshDateTime": {
                    "type": "string"
                },
                "jwtAccessToken": {
                    "type": "string"
                },
                "jwtRefreshToken": {
                    "type": "string"
                }
            }
        },
        "auth.DataUserAuthenticated": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "header_img": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nick_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "user_name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "auth.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "user_name"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "user_name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "auth.SecurityAuthenticatedUser": {
            "type": "object",
            "properties": {
                "security": {
                    "$ref": "#/definitions/auth.DataSecurityAuthenticated"
                },
                "userinfo": {
                    "$ref": "#/definitions/auth.DataUserAuthenticated"
                }
            }
        },
        "auth.SecurityRegisterUser": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/auth.DataUserAuthenticated"
                }
            }
        },
...
        "domain.CommonResponse-auth_SecurityAuthenticatedUser": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/auth.SecurityAuthenticatedUser"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "domain.CommonResponse-auth_SecurityRegisterUser": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/auth.SecurityRegisterUser"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}